Implement notifications for buyers

+++ server-dietideals24/src/main/java/com/ucm/serverdietideals24/Impl/NotificationImpl.java
@@ -23,19 +23,19 @@ public class NotificationImpl implements NotificationDAO {
 
     @Override
     public List<Notification> getAllViaUserId(Long userId) {
-        return jdbcTemplate.query("SELECT * FROM notification WHERE idUserAccountSeller = " + userId,
+        return jdbcTemplate.query("SELECT * FROM notification WHERE idUserAccount = " + userId,
                 new BeanPropertyRowMapper<Notification>(Notification.class));
     }
 
     @Override
     public void create(Notification noti) {
-        jdbcTemplate.execute("INSERT INTO notification (id, auctionname, idoffer, idauction, iduseraccountseller) VALUES('"
+        jdbcTemplate.execute("INSERT INTO notification (id, auctionname, idauction, iduseraccount) VALUES('"
                 + noti.getId() + "', '" + noti.getAuctionName()
-                + "', '0' , '" + noti.getIdAuction() + "', '" + noti.getIdUserAccountSeller() + "')");
+                + "', '" + noti.getIdAuction() + "', '" + noti.getIdUserAccount() + "')");
     }
 
     @Override
-    public void delete(Long notiId) {
-        jdbcTemplate.execute("DELETE FROM notification WHERE id = " + notiId);
+    public void delete(Long notiId, Long userId) {
+        jdbcTemplate.execute("DELETE FROM notification WHERE id = " + notiId + " AND idUserAccount = " + userId);
     }
 }
